requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
streamlit==1.26.0
matplotlib==3.8.2
pyyaml==6.0
ortools==9.7.9490
